user-service/pyproject.toml
[tool.poetry]
name = "user-service"
version = "0.1.0"
description = "User service for authentication and user management"
authors = ["Egor Muratov <egor.muratov.27@mail.ru>"]
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "0.115.6"
uvicorn = "0.34.0"
gunicorn = "23.0.0"
sqlalchemy = "2.0.36"
asyncpg = "0.30.0"
alembic = "1.14.0"
pydantic = "2.10.4"
pydantic-settings = "2.7.1"
email-validator = "^2.0.0"
python-jose = {version = "3.3.0", extras = ["cryptography"]}
bcrypt = "4.2.1"
passlib = "1.7.4"
faststream = {extras = ["kafka"], version = "^0.6.3"}
redis = "4.6.0"
dependency-injector = "^4.48.3"
prometheus-fastapi-instrumentator = "7.0.2"
sentry-sdk = "2.23.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[dependency-groups]
dev = [
    "psycopg2-binary (>=2.9.11,<3.0.0)"
]
